+++ src/TortoiseProc/RebaseDlg.cpp
@@ -291,7 +291,7 @@ BOOL CRebaseDlg::OnInitDialog()
 	if(m_CommitList.m_IsOldFirst)
 		this->m_CurrentRebaseIndex = -1;
 	else
-		this->m_CurrentRebaseIndex = m_CommitList.m_logEntries.size();
+		this->m_CurrentRebaseIndex = (int)m_CommitList.m_logEntries.size();
 
 
 	if(this->CheckRebaseCondition())
@@ -1641,7 +1641,7 @@ void CRebaseDlg::OnBnClickedRebasePostButton()
 	this->m_Upstream=this->m_UpstreamCtrl.GetString();
 	this->m_Branch=this->m_BranchCtrl.GetString();
 
-	this->EndDialog(IDC_REBASE_POST_BUTTON+this->m_PostButton.GetCurrentEntry());
+	this->EndDialog((int)IDC_REBASE_POST_BUTTON+this->m_PostButton.GetCurrentEntry());
 }
 
 void CRebaseDlg::Refresh()
